+++ tests/utils/mod.rs
@@ -2,8 +2,8 @@ pub trait IsWithin {
     fn is_within(&self, other: Self) -> bool;
 }
 
-impl IsWithin for (uint, uint) {
-    fn is_within(&self, other: (uint, uint)) -> bool {
+impl IsWithin for (usize, usize) {
+    fn is_within(&self, other: (usize, usize)) -> bool {
         self.0 <= other.0 && self.1 <= other.1
     }
 }
